# docs/security/openapi.yaml
openapi: 3.0.3
info:
  title: Civis CRM/ERP API
  version: 0.1.0
  description: |
    Generated from the current Next.js App Router endpoints under `app/api/**/route.ts`.
    This spec is intended for Swagger/SwaggerHub documentation and VAPT planning.
servers:
  - url: https://ik-crm-erp.onrender.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    nextAuthSession:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (development/non-secure cookie name).
    nextAuthSecureSession:
      type: apiKey
      in: cookie
      name: __Secure-next-auth.session-token
      description: NextAuth session cookie (secure cookie name).
  schemas:
...
    UserSettingsPatchInput:
      type: object
      properties:
        profile:
          type: object
          properties:
            name: { type: string }
            email: { type: string, format: email }
            title: { type: string }
            phone: { type: string }
            timezone: { type: string }
            locale: { type: string }
        preferences:
          type: object
          additionalProperties: true
        notifications:
          type: object
          additionalProperties: true
        kpis:
          type: array
          items:
            type: object
            additionalProperties: true
        digest:
          type: object
          additionalProperties: true
        onboarding:
          type: array
          items:
            type: object
            additionalProperties: true
        crmViews:
          type: object
          additionalProperties: true
...
security:
  - nextAuthSession: []
  - nextAuthSecureSession: []
paths:
...
  /api/user/settings:
    get:
      tags: [User Settings]
      summary: Get profile and preferences
      responses:
        '200': { description: Settings payload }
        '401': { description: Authentication required }
    patch:
      tags: [User Settings]
      summary: Update profile and preferences
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettingsPatchInput'
      responses:
        '200': { description: Settings updated }
        '400': { description: Invalid payload }
        '401': { description: Authentication required }
...